requirements.txt
Django>=3.2.4,<4.0 
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
requests>=2.25,<3.0
drf-yasg